cv/helpers: separated regression_errors.py from prior_comparison_errors.py updated semisynthetic and semisynthetic2

+++ scripts/figures/keystoneness.py
@@ -121,7 +121,7 @@ class MdsineOutput(object):
                 len(clusters),
                 len(clusters)
             ),
-            dtype=np.float
+            dtype=float
         )
         cluster_reps = [
             next(iter(cluster.members)) for cluster in clusters
